Add better Java docs
