+++ src/main/java/io/auklet/config/AbstractConfigFileFromApi.java
@@ -3,7 +3,7 @@ package io.auklet.config;
 import edu.umd.cs.findbugs.annotations.CheckForNull;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import io.auklet.AukletException;
-import io.auklet.core.Util;
+import io.auklet.misc.Util;
 import net.jcip.annotations.NotThreadSafe;
 
 import java.io.IOException;
