+++ src/main/java/io/auklet/config/AbstractJsonConfigFileFromApi.java
@@ -2,7 +2,7 @@ package io.auklet.config;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import io.auklet.AukletException;
-import io.auklet.core.Util;
+import io.auklet.misc.Util;
 import mjson.Json;
 import net.jcip.annotations.NotThreadSafe;
 import okhttp3.Request;
