+++ src/main/java/io/auklet/config/DataUsageLimit.java
@@ -4,7 +4,7 @@ import edu.umd.cs.findbugs.annotations.NonNull;
 import io.auklet.Auklet;
 import io.auklet.AukletException;
 import io.auklet.core.DataUsageConfig;
-import io.auklet.core.Util;
+import io.auklet.misc.Util;
 import mjson.Json;
 import net.jcip.annotations.NotThreadSafe;
 import okhttp3.Request;
