--- src/main/java/io/auklet/jvm/package-info.java
@@ -1,2 +1,0 @@
-/** <p>Classes that provide integrations specific to JVM implementations.</p> */
-package io.auklet.jvm;
