+++ src/main/java/io/auklet/metrics/package-info.java
@@ -1,0 +1,2 @@
+/** <p>Classes that provide helper functions to retrieve metrics from different JVM Platform.</p> */
+package io.auklet.metrics;
