+++ src/main/java/io/auklet/misc/package-info.java
@@ -1,0 +1,2 @@
+/** <p>Classes that provide helper miscellaneous functions.</p> */
+package io.auklet.misc;
