+++ src/main/java/io/auklet/sink/AbstractSink.java
@@ -5,7 +5,7 @@ import edu.umd.cs.findbugs.annotations.Nullable;
 import io.auklet.Auklet;
 import io.auklet.AukletException;
 import io.auklet.core.HasAgent;
-import io.auklet.core.Util;
+import io.auklet.misc.Util;
 import net.jcip.annotations.ThreadSafe;
 import org.msgpack.core.MessageBufferPacker;
 import org.msgpack.core.MessagePack;
